backend/requirements.txt
# Core
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
elasticsearch==8.12.0
elasticsearch[async]==8.12.0

# SQL Database (metadados do sistema)
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
# psycopg2-binary==2.9.9  # Removed - usando asyncpg instead

# Encryption (para senhas de ES servers)
cryptography==41.0.7

# Cache & Pub/Sub (opcional)
redis==5.0.1

# WebSocket
python-socketio==5.10.0
python-engineio==4.8.0

# AI & LLM
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.23
anthropic==0.21.3
openai==1.54.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dateutil>=2.9.0
httpx==0.25.2

# RSS Feed Processing
feedparser==6.0.11
pyyaml==6.0.1
celery==5.3.4
requests>=2.32.5

# Web Scraping (Malpedia sync)
beautifulsoup4==4.12.2
lxml==4.9.3

# SSO & OAuth2
authlib==1.3.0
msal==1.28.0
apscheduler==3.10.4

# Export (PDF & Images)
reportlab==4.0.7
playwright==1.40.0
pillow==11.0.0  # Updated for Python 3.13 support
plotly==5.18.0
kaleido==0.2.1

# CTI & Threat Intelligence
mitreattack-python==3.0.3
stix2==3.0.1
pymisp>=2.5.17
OTXv2>=1.5.12

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
OTXv2==1.5.12
